eduverse/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
google-genai
reportlab==4.0.0
python-multipart
passlib[bcrypt]
python-jose[cryptography]
